requirements.txt
aiosmtplib==4.0.2
alembic==1.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
bcrypt==3.2.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.1.8
cryptography==46.0.5
Deprecated==1.3.1
ecdsa==0.19.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.128.8
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.1.0
joblib==1.5.3
limits==4.2
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.0.2
openpyxl==3.1.5
packaging==24.2
passlib==1.7.4
pillow==11.3.0
pluggy==1.6.0
pyasn1==0.6.3
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyMySQL==1.1.2
pytest==8.4.2
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
reportlab==4.4.10
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.48
starlette==0.49.3
threadpoolctl==3.6.0
tomli==2.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uvicorn==0.39.0
websockets==15.0.1
wrapt==2.1.2
